Rename files under src/video/qnx
